save mapconfig
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Reporting/APP@6976 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/reporting/service/bean/SpatialServiceBean.java
@@ -52,8 +52,8 @@ public class SpatialServiceBean implements SpatialService {
         try {
             validate(mapConfiguration);
 
-            Integer mapProjection = mapConfiguration.getMapProjection();
-            Integer displayProjection = mapConfiguration.getDisplayProjection();
+            Long mapProjection = Long.valueOf(mapConfiguration.getMapProjection());
+            Long displayProjection = Long.valueOf(mapConfiguration.getDisplayProjection());
             CoordinatesFormat coordinatesFormat = null;
             if (mapConfiguration.getCoordinatesFormat() != null) {
                 coordinatesFormat = CoordinatesFormat.fromValue(mapConfiguration.getCoordinatesFormat().toUpperCase());
@@ -76,7 +76,7 @@ public class SpatialServiceBean implements SpatialService {
         }
     }
 
-    private String getSaveMapConfigurationRequest(Long reportId, Integer mapProjection, Integer displayProjection, CoordinatesFormat coordinatesFormat, ScaleBarUnits scaleBarUnits) throws SpatialModelMarshallException {
+    private String getSaveMapConfigurationRequest(Long reportId, Long mapProjection, Long displayProjection, CoordinatesFormat coordinatesFormat, ScaleBarUnits scaleBarUnits) throws SpatialModelMarshallException {
         return SpatialModuleRequestMapper.mapToSpatialSaveMapConfigurationRQ(reportId, mapProjection, displayProjection, coordinatesFormat, scaleBarUnits);
     }
